refactor: URL 경로 수정 및 PAGE_URL 상수 폴더 변경 #34
